Construction functions extracted
